Seperate Styles Relates #79
