+++ client/src/pages/errors/404.jsx
@@ -1,14 +1,7 @@
-const styles = {
-  container: {
-    display: 'flex',
-    flexDirection: 'column',
-    alignItems: 'center',
-    justifyContent: 'center',
-    height: '100vh',
-  },
-};
+import styles from './styles';
+
 const Error404 = () => (
-  <div style={styles.container}>
+  <div style={styles.clientContainer}>
     <h1> عفواً ... خطأ 404   </h1>
     <h2>عذرا، ولكن الصفحة التي تبحث عنها غير موجودة </h2>
     <p>
