+++ client/src/pages/errors/500.jsx
@@ -1,15 +1,7 @@
-const styles = {
-  container: {
-    display: 'flex',
-    flexDirection: 'column',
-    alignItems: 'center',
-    justifyContent: 'center',
-    height: '100vh',
-  },
-};
+import styles from './styles';
 
 const Error500 = () => (
-  <div style={styles.container}>
+  <div style={styles.serverContainer}>
     <h1> خطأ 500 </h1>
     <h3>نأسف لقد حدث خطأ ما خلال معالجة بعض البيانات</h3>
     <p>
